Added 3 new enhancements
(No 1 Enhancement )
Login and Register- I made the login and register with database system 

(No 2 Enhancement)
Program will automatically share transfer details with email- After the user transfers the money to other people, 
the user can easily share this transfer details with email by entering the receiver email address.

(No 3 Enhancement)
Enhanced the old "check_transactions" def - I enhanced it because the old function is complicated and hard to read. For instance it show the user like this 

{"index": 14, "timestamp": "1601195086.559453", "data": {"proof-of-work": 583852032, "transactions": [{"from": "WYLMvvTYkqxVYpWq2y7CGyz6aCO4DVkeeDrJx59FdApUhRNNV+D5iItr5wOB0iY5KYQ5fCDERg4xLw0Cu/fAjg==", "to": "eNcAW03xvdsCi9t4pH+ZwU+d1lirROMuDpxelnT5O4nnknUJBo50tDkz3RP5e3o17daExRwc2YdEXuy2W7D/kg==", "amount": "2200", "signature": "AiOG/Io9dPeBW384pPOGL4wWcxIbUUE+R5P1ZdjO7eXqJYHx+a+TWaXfkBgHojmRHsM6PEelQI6U8fwHj5cAcg==", "message": "1601195020"}, {"from": "network", "to": "q3nf394hjg-random-miner-address-34nf3i4nflkn3oi", "amount": 1}]}, "hash": "f1911b9b1bc99a002c54f9416401f5550c6dd9e4e8924318c68fa24b83e7462f"} {"index": 14, "timestamp": "1601195086.559453", "data": {"proof-of-work": 583852032, "transactions": [{"from": "WYLMvvTYkqxVYpWq2y7CGyz6aCO4DVkeeDrJx59FdApUhRNNV+D5iItr5wOB0iY5KYQ5fCDERg4xLw0Cu/fAjg==", "to": "eNcAW03xvdsCi9t4pH+ZwU+d1lirROMuDpxelnT5O4nnknUJBo50tDkz3RP5e3o17daExRwc2YdEXuy2W7D/kg==", "amount": "2200", "signature": "AiOG/Io9dPeBW384pPOGL4wWcxIbUUE+R5P1ZdjO7eXqJYHx+a+TWaXfkBgHojmRHsM6PEelQI6U8fwHj5cAcg==", "message": "1601195020"}, {"from": "network", "to": "q3nf394hjg-random-miner-address-34nf3i4nflkn3oi", "amount": 1}]}, "hash": "f1911b9b1bc99a002c54f9416401f5550c6dd9e4e8924318c68fa24b83e7462f"}

This history also disappeared when the user closed the program. This is why I created a database and stored every transaction in db. Users can easily retrieve their transaction history from the database and check them. 
Moreover, I also added the data and time log in this.